+++ packages/core/src/singleton/AmplifyOutputs/types.ts
@@ -42,7 +42,7 @@ export interface AmplifyOutputsAuthProperties {
 	unauthenticated_identities_enabled?: boolean;
 	mfa_configuration?: string;
 	mfa_methods?: string[];
-	groups?: Record<UserGroupName, UserGroupPrecedence>[];
+	groups?: Partial<Record<UserGroupName, UserGroupPrecedence>>[];
 }
 
 export interface AmplifyOutputsStorageBucketProperties {
@@ -53,7 +53,7 @@ export interface AmplifyOutputsStorageBucketProperties {
 	/** Region for the bucket */
 	aws_region: string;
 	/** Paths to object with access permissions */
-	paths?: Record<string, Record<string, string[] | undefined>>;
+	paths?: Partial<Record<string, Record<string, string[] | undefined>>>;
 }
 export interface AmplifyOutputsStorageProperties {
 	/** Default region for Storage */